reorg, lots of docstrings for analysis test utils
